Add part of catenary.
