--- doc/modules/canvas.rst
@@ -1,18 +1,0 @@
-pyglet.canvas
-=============
-
-.. automodule:: pyglet.canvas
-  :members:
-  :undoc-members:
-
-.. autoclass:: Display
-  :members:
-  :undoc-members:
-
-.. autoclass:: Screen
-  :members:
-  :undoc-members:
-
-.. autoclass:: Canvas
-  :members:
-  :undoc-members:
